implemented review suggestions
